Bug 330135 - Start e4 RCP several times fails
The review comments have been applied

+++ bundles/org.eclipse.e4.ui.workbench/src/org/eclipse/e4/ui/internal/workbench/ResourceHandler.java
@@ -14,11 +14,10 @@
 
 package org.eclipse.e4.ui.internal.workbench;
 
-import org.eclipse.e4.ui.workbench.IWorkbench;
-
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.net.URLConnection;
 import java.util.Collection;
 import java.util.Collections;
@@ -31,6 +30,7 @@ import org.eclipse.core.internal.runtime.PlatformURLPluginConnection;
 import org.eclipse.core.runtime.URIUtil;
 import org.eclipse.e4.core.contexts.ContextInjectionFactory;
 import org.eclipse.e4.core.contexts.IEclipseContext;
+import org.eclipse.e4.core.di.annotations.Optional;
 import org.eclipse.e4.core.services.log.Logger;
 import org.eclipse.e4.ui.model.application.MApplication;
 import org.eclipse.e4.ui.model.application.MApplicationElement;
@@ -41,6 +41,7 @@ import org.eclipse.e4.ui.model.application.ui.basic.impl.BasicPackageImpl;
 import org.eclipse.e4.ui.model.application.ui.impl.UiPackageImpl;
 import org.eclipse.e4.ui.model.application.ui.menu.impl.MenuPackageImpl;
 import org.eclipse.e4.ui.workbench.IModelResourceHandler;
+import org.eclipse.e4.ui.workbench.IWorkbench;
 import org.eclipse.e4.ui.workbench.modeling.IModelReconcilingService;
 import org.eclipse.e4.ui.workbench.modeling.ModelDelta;
 import org.eclipse.e4.ui.workbench.modeling.ModelReconciler;
@@ -74,6 +75,7 @@ public class ResourceHandler implements IModelResourceHandler {
 	private URI applicationDefinitionInstance;
 
 	@Inject
+	@Optional
 	@Named(E4Workbench.INSTANCE_LOCATION)
 	private Location instanceLocation;
 
@@ -127,10 +129,11 @@ public class ResourceHandler implements IModelResourceHandler {
 	}
 
 	public Resource loadMostRecentModel() {
-		File baseLocation = getBaseLocation();
 		// This is temporary code to migrate existing delta files into full models
 		if (deltaRestore && saveAndRestore && !clearPersistedState) {
+			File baseLocation = getBaseLocation();
 			File deltaFile = new File(baseLocation, "deltas.xml"); //$NON-NLS-1$
+
 			if (deltaFile.exists()) {
 				MApplication appElement = null;
 				try {
@@ -175,14 +178,17 @@ public class ResourceHandler implements IModelResourceHandler {
 			}
 		}
 
-		File workbenchData = getWorkbenchSaveLocation();
-
-		if (clearPersistedState && workbenchData.exists())
-			workbenchData.delete();
-
+		File workbenchData = null;
 		URI restoreLocation = null;
-		if (saveAndRestore)
+
+		if (saveAndRestore) {
+			workbenchData = getWorkbenchSaveLocation();
 			restoreLocation = URI.createFileURI(workbenchData.getAbsolutePath());
+		}
+
+		if (clearPersistedState && workbenchData != null && workbenchData.exists()) {
+			workbenchData.delete();
+		}
 
 		// last stored time-stamp
 		long restoreLastModified = restoreLocation == null ? 0L : new File(
@@ -228,10 +234,17 @@ public class ResourceHandler implements IModelResourceHandler {
 	 * @return a resource with a proper save path with the model as contents
 	 */
 	public Resource createResourceWithApp(MApplication theApp) {
-		URI saveLocation = URI.createFileURI(getWorkbenchSaveLocation().getAbsolutePath());
-		Resource res = resourceSetImpl.createResource(saveLocation);
+		Resource res = createResource();
 		res.getContents().add((EObject) theApp);
 		return res;
+	}
+
+	private Resource createResource() {
+		if (saveAndRestore) {
+			URI saveLocation = URI.createFileURI(getWorkbenchSaveLocation().getAbsolutePath());
+			return resourceSetImpl.createResource(saveLocation);
+		}
+		return resourceSetImpl.createResource(URI.createURI("workbench.xmi")); //$NON-NLS-1$
 	}
 
 	private File getWorkbenchSaveLocation() {
@@ -256,7 +269,7 @@ public class ResourceHandler implements IModelResourceHandler {
 	private Resource loadResource(URI uri) {
 		Resource resource;
 		try {
-			resource = resourceSetImpl.getResource(uri, true);
+			resource = getResource(uri);
 		} catch (Exception e) {
 			// TODO We could use diagnostics for better analyzing the error
 			logger.error(e, "Unable to load resource " + uri.toString()); //$NON-NLS-1$
@@ -274,6 +287,22 @@ public class ResourceHandler implements IModelResourceHandler {
 				}
 			}
 		}
+		return resource;
+	}
+
+	private Resource getResource(URI uri) throws Exception {
+		Resource resource;
+		if (saveAndRestore) {
+			resource = resourceSetImpl.getResource(uri, true);
+		} else {
+			// Workaround for java.lang.IllegalStateException: No instance data can be specified
+			// thrown by org.eclipse.core.internal.runtime.DataArea.assertLocationInitialized
+			// The DataArea.assertLocationInitialized is called by ResourceSetImpl.getResource(URI,
+			// boolean)
+			resource = resourceSetImpl.createResource(uri);
+			resource.load(new URL(uri.toString()).openStream(), resourceSetImpl.getLoadOptions());
+		}
+
 		return resource;
 	}
